move autosubscribe logic to AllServicesOn screen, fix navigation on banner
